RadioGroup is now exported as Radio.Group, no longer in two areas in the docs
