+++ src/components/molecules/form/field/field.md
@@ -8,6 +8,7 @@ Form Fields are a group of field types supported in `Form`:
 * `Form.TextArea`
 * `Form.Select`
 * `Form.Switch`
+* `Form.RadioGroup`
 
 In addition to their own [native props](/docs/TextInput), we add a few more props in the context of a `Form`:
