+++ src/components/molecules/form/index.js
@@ -5,7 +5,7 @@ import ActionInput from '../../molecules/_action-input'
 import TextArea from '../../atoms/textarea'
 import Select from '../../atoms/select'
 import Switch from '../../atoms/switch'
-import RadioGroup from '../../molecules/radio-group'
+import Radio from '../../atoms/radio'
 
 import Field from './field'
 import Actions from './actions'
@@ -23,7 +23,7 @@ Form.TextInput = props => <Field {...props} fieldComponent={ActionInput} />
 Form.TextArea = props => <Field {...props} fieldComponent={TextArea} />
 Form.Select = props => <Field {...props} fieldComponent={Select} />
 Form.Switch = props => <Field {...props} fieldComponent={Switch} />
-Form.RadioGroup = props => <Field {...props} fieldComponent={RadioGroup} />
+Form.RadioGroup = props => <Field {...props} fieldComponent={Radio.Group} />
 Form.Actions = Actions
 Form.FieldSet = FieldSet
