--- src/components/molecules/radio-group/radio-group.md
@@ -1,15 +1,0 @@
-```meta
-  category: Layout
-  description: Use a RadioGroup to put related radios together
-```
-
-`import { RadioGroup } from '@auth0/cosmos'`
-
----
-
-```jsx
-<RadioGroup {props} name="opt">
-  <Radio value="opt1">Option 1</Radio>
-  <Radio value="opt2">Option 2</Radio>
-</RadioGroup>
-```
